# docs/openapi.yaml
openapi: 3.1.0
info:
  title: BizEngine API
  version: 1.0.0
  description: Business engine for managing organizations, products, orders, warehouses, HR, finance, logistics, and Russian government integrations.
servers:
  - url: http://localhost:8080
    description: Local development
  - url: https://api.bizengine.app
    description: Production
...
paths:
...
  /api/v1/organizations/{orgID}/white-label:
    get:
      tags: [Banking]
      summary: Get white-label banking configuration
      operationId: getWhiteLabel
      parameters:
        - $ref: '#/components/parameters/orgID'
      responses:
        '200':
          description: White-label config
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
                  data:
                    type: object
                    properties:
                      bank_name: { type: string }
                      bank_code: { type: string }
                      connected: { type: boolean }
                      features: { type: array, items: { type: string } }
  /api/v1/organizations/{orgID}/payments/initiate:
    post:
      tags: [Banking]
      summary: Initiate a payment
      operationId: initiatePayment
      parameters:
        - $ref: '#/components/parameters/orgID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [data]
              properties:
                data:
                  type: object
                  required: [recipient_account, recipient_bik, recipient_name, recipient_inn, amount, currency, purpose]
                  properties:
                    recipient_account: { type: string }
                    recipient_bik: { type: string }
                    recipient_name: { type: string }
                    recipient_inn: { type: string }
                    amount: { type: integer, description: Amount in minor units (kopecks) }
                    currency: { type: string, enum: [RUB] }
                    purpose: { type: string }
      responses:
        '201':
          description: Payment initiated
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
                  data:
                    type: object
                    properties:
                      payment_id: { type: string, format: uuid }
                      status: { type: string, enum: [pending, processing, completed, failed] }
                      created_at: { type: string, format: date-time }
  /api/v1/organizations/{orgID}/payments/{paymentID}/status:
    get:
      tags: [Banking]
      summary: Check payment status
      operationId: getPaymentStatus
      parameters:
        - $ref: '#/components/parameters/orgID'
        - name: paymentID
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Payment status
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
                  data:
                    type: object
                    properties:
                      payment_id: { type: string, format: uuid }
                      status: { type: string, enum: [pending, processing, completed, failed] }
                      amount: { type: integer }
                      currency: { type: string }
                      created_at: { type: string, format: date-time }
                      updated_at: { type: string, format: date-time }
                      error_message: { type: string }
  /api/v1/organizations/{orgID}/finance/bank/auto-reconcile:
    post:
      tags: [Banking]
      summary: Auto-reconcile bank transactions
      operationId: autoReconcile
      parameters:
        - $ref: '#/components/parameters/orgID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [data]
              properties:
                data:
                  type: object
                  required: [bank_account_id, date_from, date_to]
                  properties:
                    bank_account_id: { type: string, format: uuid }
                    date_from: { type: string, format: date }
                    date_to: { type: string, format: date }
      responses:
        '200':
          description: Reconciliation result
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
                  data:
                    type: object
                    properties:
                      reconciliation_id: { type: string, format: uuid }
                      matched: { type: integer }
                      unmatched: { type: integer }
                      auto_matched: { type: integer }
                      manual_review: { type: integer }
                      status: { type: string }
  /api/v1/organizations/{orgID}/finance/bank/reconciliations:
    get:
      tags: [Banking]
      summary: List reconciliation runs
      operationId: listReconciliations
      parameters:
        - $ref: '#/components/parameters/orgID'
      responses:
        '200':
          description: Reconciliation list
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        id: { type: string, format: uuid }
                        bank_account_id: { type: string, format: uuid }
                        date_from: { type: string, format: date }
                        date_to: { type: string, format: date }
                        matched: { type: integer }
                        unmatched: { type: integer }
                        status: { type: string }
                        created_at: { type: string, format: date-time }
  /api/v1/organizations/{orgID}/finance/bank/reconciliations/{id}:
    get:
      tags: [Banking]
      summary: Get reconciliation details
      operationId: getReconciliation
      parameters:
        - $ref: '#/components/parameters/orgID'
        - $ref: '#/components/parameters/id'
      responses:
        '200':
          description: Reconciliation details
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
                  data:
                    type: object
                    properties:
                      id: { type: string, format: uuid }
                      bank_account_id: { type: string, format: uuid }
                      date_from: { type: string, format: date }
                      date_to: { type: string, format: date }
                      matched: { type: integer }
                      unmatched: { type: integer }
                      auto_matched: { type: integer }
                      manual_review: { type: integer }
                      status: { type: string }
                      entries:
                        type: array
                        items:
                          type: object
                          properties:
                            bank_transaction_id: { type: string }
                            matched_transaction_id: { type: string, format: uuid }
                            match_type: { type: string, enum: [auto, manual, unmatched] }
                            amount: { type: integer }
...
components:
  parameters:
    orgID:
      name: orgID
      in: path
      required: true
      schema:
        type: string
        format: uuid
    id:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid